Flask/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
Flask-RESTful==0.3.9
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
requests
flask-socketio
eventlet
pysmb
